refactro: extract helper functions to own file
