+++ packages/shared/lib/core/wallet/utils/generateActivity/helper/index.ts
@@ -2,6 +2,7 @@ export * from './getActivityTypeFromOutput'
 export * from './getAsyncDataFromOutput'
 export * from './getAmountFromOutput'
 export * from './getSendingInformation'
+export * from './getGovernanceInfo'
 export * from './getGovernorAddressFromAliasOutput'
 export * from './getMetadataFromOutput'
 export * from './getStateControllerAddressFromAliasOutput'
